Make `submit` a method of `Submitter`.
